requirements.txt
boto3>=1.28.0
numpy>=1.24.0
Pillow>=10.0.0
